fix iteration four add more recipes
Added two more recipes with identical page structures with a few additions to the recipe page

the recipes are spinach quiche and italian sausage peppers and onions.
created the respective html pages
wrote headings and tagged images
linked the pages to index.html
fixed the image paths so that the images can be seen on the site.